requirements.txt
pandas
numpy
matplotlib
scikit-learn
fastapi
uvicorn
pydantic
ipykernel
supabase
psutil
speedtest-cli
requests
beautifulsoup4
python-dotenv
selenium